Modify library references
As described in the README of the newly added libraries submodule.
Schematic and PCB are synced.

Hardware folder is renamed to mainboard
Any other PCBs can be put in their own subfolder in the gitrepo

Commit also swaps source/drain of Q1 because it appeared wrong for some
reason after updating all symbol/footprints.